Fields

Fields

Oceananigans.CellField — Type.

CellField

A field defined at the cell centers. Used for pressure and tracers.

source

Oceananigans.FaceFieldX — Type.

FaceFieldX

A field defined at the faces along the x-direction. Used for horizontal velocity u.

source

Oceananigans.FaceFieldY — Type.

FaceFieldY

A field defined at the faces along the y-direction. Used for horizontal velocity v.

source

Oceananigans.FaceFieldZ — Type.

FaceFieldY

A field defined at the faces along the z-direction. Used for vertical velocity w.

source

Missing docstring.

Missing docstring for EdgeField. Check Documenter's build log for details.