Grids

Grids

Oceananigans.RegularCartesianGrid — Type.

RegularCartesianGrid{T<:AbstractFloat, R<:AbstractRange} <: Grid

A Cartesian grid with regularly spaces cells and faces so that $Δx$, $Δy$, and $Δz$ are constants. Fields are stored using floating-point values of type T. The cell-centered and face-centered coordinate ranges are stored as ranges of type A.

source