Fields

Fields

Oceananigans.CellField — Type.

CellField{A<:AbstractArray, G<:Grid} <: Field

A cell-centered field defined on a grid G whose values are stored in an A.

source

Oceananigans.FaceFieldX — Type.

FaceFieldX{A<:AbstractArray, G<:Grid} <: FaceField

An x-face-centered field defined on a grid G whose values are stored in an A.

source

Oceananigans.FaceFieldY — Type.

FaceFieldY{A<:AbstractArray, G<:Grid} <: FaceField

A y-face-centered field defined on a grid G whose values are stored in an A.

source

Oceananigans.FaceFieldZ — Type.

FaceFieldZ{T, G<:Grid{T}} <: FaceField{G}

A z-face-centered field defined on a grid G whose values are stored in an A.

source

Oceananigans.EdgeField — Type.

EdgeField{T<:AbstractArray} <: Field

An edge-centered field defined on a grid G whose values are stored in an A.

source