Fields

Fields

Oceananigans.CellField — Type.

CellField{T,G<:Grid{T}} <: Field{G}

A cell-centered field defined on a grid G whose values are stored as floating-point values of type T.

source

Oceananigans.FaceFieldX — Type.

FaceFieldX{T,G<:Grid{T}} <: FaceField{G}

An x-face-centered field defined on a grid G whose values are stored as floating-point values of type T.

source

Oceananigans.FaceFieldY — Type.

FaceFieldY{T,G<:Grid{T}} <: FaceField{G}

A y-face-centered field defined on a grid G whose values are stored as floating-point values of type T.

source

Oceananigans.FaceFieldZ — Type.

FaceFieldZ{T,G<:Grid{T}} <: FaceField{G}

A z-face-centered field defined on a grid G whose values are stored as floating-point values of type T.

source

Oceananigans.EdgeField — Type.

EdgeField{T<:AbstractArray} <: Field

A field defined on a grid G whose values lie on the edges of the cells.

source

Oceananigans.CellField — Method.

CellField(metadata::ModelMetadata, grid::Grid, T)

Construct a CellField whose values are defined at the center of a cell.

source

Oceananigans.FaceFieldX — Method.

FaceFieldX(metadata::ModelMetadata, grid::Grid, T)

A Field whose values are defined on the x-face of a cell.

source

Oceananigans.FaceFieldY — Method.

FaceFieldY(metadata::ModelMetadata, grid::Grid, T)

A Field whose values are defined on the y-face of a cell.

source

Oceananigans.FaceFieldZ — Method.

FaceFieldZ(metadata::ModelMetadata, grid::Grid, T)

A Field whose values are defined on the z-face of a cell.

source

Oceananigans.EdgeField — Method.

FEdgeField(metadata::ModelMetadata, grid::Grid, T)

A Field whose values are defined on the edges of a cell.

source