Model setup

This section describes all the options and features that can be used to set up a model. For more detailed information consult the API documentation.

Each structure covered in this section can be constructed and passed to the IncompressibleModel constructor. For examples of model construction, see the examples. The verification experiments provide more advanced examples.

For reference, here are all the option or keyword arguments that can be passed to IncompressibleModel. See the different sections on the sidebar for more details and examples for each keyword argument.

Oceananigans.Models.IncompressibleModel — Type

IncompressibleModel(;
               grid,
       architecture = CPU(),
         float_type = Float64,
              clock = Clock{float_type}(0, 0, 1),
          advection = CenteredSecondOrder(),
           buoyancy = SeawaterBuoyancy(float_type),
           coriolis = nothing,
      surface_waves = nothing,
            forcing = NamedTuple(),
            closure = IsotropicDiffusivity(float_type, ν=ν₀, κ=κ₀),
boundary_conditions = NamedTuple(),
            tracers = (:T, :S),
        timestepper = :QuasiAdamsBashforth2,
         velocities = nothing,
          pressures = nothing,
      diffusivities = nothing,
    pressure_solver = nothing
)

Construct an incompressible Oceananigans.jl model on grid.

Keyword arguments

- `grid`: (required) The resolution and discrete geometry on which `model` is solved.
- `architecture`: `CPU()` or `GPU()`. The computer architecture used to time-step `model`.
- `float_type`: `Float32` or `Float64`. The floating point type used for `model` data.
- `advection`: The scheme that advects velocities and tracers. See `Oceananigans.Advection`.
- `buoyancy`: The buoyancy model. See `Oceananigans.Buoyancy`.
- `closure`: The turbulence closure for `model`. See `Oceananigans.TurbulenceClosures`.
- `coriolis`: Parameters for the background rotation rate of the model.
- `forcing`: `NamedTuple` of user-defined forcing functions that contribute to solution tendencies.
- `boundary_conditions`: `NamedTuple` containing field boundary conditions.
- `tracers`: A tuple of symbols defining the names of the modeled tracers, or a `NamedTuple` of
             preallocated `CellField`s.
- `timestepper`: A symbol that specifies the time-stepping method. Either `:QuasiAdamsBashforth2` or
                 `:RungeKutta3`.